+++ src/model/PizzaSize.java
@@ -1,17 +1,17 @@
 package model;
 
 public enum PizzaSize {
-	SMALL(7.00, "SMALL"), MEDIUM(9.00, "MEDIUM"), LARGE(11.00, "LARGE"), EXTRALARGE(13.00, "EXTRA LARGE");
+	SMALL((long)7.00, "SMALL"), MEDIUM((long)9.00, "MEDIUM"), LARGE((long)11.00, "LARGE"), EXTRALARGE((long)13.00, "EXTRA LARGE");
 	
-	private final double cost;
+	private final long cost;
 	private final String desc;
 	
-	PizzaSize(double cost, String desc){
+	PizzaSize(long cost, String desc){
 		this.cost = cost;
 		this.desc = desc;
 	}
 	
-	public double getCost(){
+	public long getCost(){
 		return this.cost;
 	}
